# Use encoding/binary instead of unsafe (fixes #29)